+++ back-end/lib/db.js
@@ -64,6 +64,13 @@ module.exports = {
       );
       return merge(message, { channelId: channelId, creation: creation });
     },
+    update: async (channelId, message) => {
+      if (!channelId) throw Error("Missing channel");
+      if (!message.author) throw Error("Missing author");
+      if (!message.content) throw Error("Missing content");
+      await db.put(`messages:${channelId}:${message.creation}`, JSON.stringify(message));
+      return merge(message, { channelId: channelId, creation: message.creation });
+    },
     list: async (channelId) => {
       return new Promise((resolve, reject) => {
         const messages = [];