+++ front-end/src/channel/List.js
@@ -22,7 +22,7 @@ import relativeTime from "dayjs/plugin/relativeTime";
 import Button from "@material-ui/core/Button";
 import TextField from "@material-ui/core/TextField";
 import Grid from "@material-ui/core/Grid";
-import { orange } from "@material-ui/core/colors";
+import { grey, orange } from "@material-ui/core/colors";
 import axios from "axios";
 import Dialog from "@material-ui/core/Dialog";
 import DialogActions from "@material-ui/core/DialogActions";
@@ -36,6 +36,7 @@ import PersonIcon from "@material-ui/icons/Person";
 import AddIcon from "@material-ui/icons/Add";
 import Avatar from "@material-ui/core/Avatar";
 import Context from "../Context";
+import Link from "@material-ui/core/Link";
 
 const useStyles = (theme) => ({
   root: {
@@ -74,6 +75,9 @@ const useStyles = (theme) => ({
     top: 0,
     width: "50px",
   },
+  infos: {
+    color: "rgb(180, 180, 180)",
+  },
 });
 
 const ColorButton = withStyles((theme) => ({
@@ -108,7 +112,14 @@ const ColorTextField = withStyles((theme) => ({
   },
 }))(TextField);
 
-dayjs.extend(relativeTime)
+const ColorLink = withStyles((theme) => ({
+  root: {
+    color: "rgb(200, 125, 0)",
+    cursor: "pointer",
+  },
+}))(Link);
+
+dayjs.extend(relativeTime);
 
 export default forwardRef(({ channel, messages, onScrollDown }, ref) => {
   const styles = useStyles(useTheme());
@@ -153,6 +164,11 @@ export default forwardRef(({ channel, messages, onScrollDown }, ref) => {
 
   const [inviteEmail, setInviteEmail] = useState("");
 
+  const [modifyOpen, setModifyOpen] = useState(false);
+  const [deleteOpen, setDeleteOpen] = useState(false);
+  const [editing, setEditing] = useState({});
+  const [newMessage, setNewMessage] = useState("");
+
   useEffect(() => {
     const getUsers = async () => {
       const queryString = window.location.pathname;
@@ -162,9 +178,9 @@ export default forwardRef(({ channel, messages, onScrollDown }, ref) => {
         `http://localhost:3001/channels/${channelID}`,
         {
           headers: {
-            'Authorization': `Bearer ${oauth.access_token}`,
-            'email': oauth.email
-          }
+            Authorization: `Bearer ${oauth.access_token}`,
+            email: oauth.email,
+          },
         }
       );
 
@@ -189,9 +205,9 @@ export default forwardRef(({ channel, messages, onScrollDown }, ref) => {
         },
         {
           headers: {
-            'Authorization': `Bearer ${oauth.access_token}`,
-            'email': oauth.email
-          }
+            Authorization: `Bearer ${oauth.access_token}`,
+            email: oauth.email,
+          },
         }
       );
 
@@ -237,9 +253,37 @@ export default forwardRef(({ channel, messages, onScrollDown }, ref) => {
           return (
             <li key={i} css={styles.message}>
               <p>
-                <span>{message.author}</span>
-                {" - "}
-                <span>{dayjs(message.creation * 1/1000).fromNow()}</span>
+                <span css={styles.infos}>
+                  {message.author}
+                  {" - "}
+                  {dayjs((message.creation * 1) / 1000).fromNow()}
+
+                  {message.author === oauth.email ? (
+                    <span>
+                      {" - ("}
+                      <ColorLink
+                        onClick={() => {
+                          setModifyOpen(true);
+                          setEditing(message);
+                        }}
+                      >
+                        modify
+                      </ColorLink>
+                      {", "}
+                      <ColorLink
+                        onClick={() => {
+                          setDeleteOpen(true);
+                          setEditing(message.id);
+                        }}
+                      >
+                        delete
+                      </ColorLink>
+                      {")"}
+                    </span>
+                  ) : (
+                    <span></span>
+                  )}
+                </span>
               </p>
               <div dangerouslySetInnerHTML={{ __html: content }}></div>
             </li>
@@ -349,6 +393,76 @@ export default forwardRef(({ channel, messages, onScrollDown }, ref) => {
             }
             <ColorButton type="submit" color="primary">
               Invite member
+            </ColorButton>
+          </DialogActions>
+        </form>
+      </Dialog>
+
+      {/* Modifying message dialog */}
+      <Dialog
+        open={modifyOpen}
+        onClose={() => {
+          setEditing(null);
+          setModifyOpen(false);
+        }}
+        aria-labelledby="form-dialog-title"
+        fullWidth
+        maxWidth="md"
+      >
+        <form
+          onSubmit={async (e) => {
+            e.preventDefault();
+            console.log(newMessage);
+            /* update message in DB */
+            const { data: result } = await axios.put(
+              `http://localhost:3001/channels/${channelData.id}/messages`,
+              {
+                content: newMessage,
+                author: editing.author,
+                channelId: editing.channelId,
+                creation: editing.creation,
+              },
+              {
+                headers: {
+                  Authorization: `Bearer ${oauth.access_token}`,
+                  email: oauth.email,
+                },
+              }
+            );
+            /* close dialog */
+            setEditing(null);
+            setNewMessage("");
+            setModifyOpen(false);
+            window.location.reload(false);
+          }}
+        >
+          <DialogContent>
+            <ColorTextField
+              autoFocus
+              margin="dense"
+              id="newMessage"
+              label="New message:"
+              type="text"
+              fullWidth
+              value={newMessage}
+              onChange={(e) => setNewMessage(e.target.value)}
+              required
+            />
+          </DialogContent>
+          <DialogActions>
+            {
+              <Button
+                onClick={() => {
+                  setEditing(null);
+                  setModifyOpen(false);
+                }}
+                color="error"
+              >
+                Cancel
+              </Button>
+            }
+            <ColorButton type="submit" color="primary">
+              Modify message
             </ColorButton>
           </DialogActions>
         </form>